+++ manifest/scripts/utils.ts
@@ -1,19 +1,12 @@
-/*
- * @Author: Cphayim
- * @Date: 2021-07-17 13:40:26
- * @Description: generate vrn-manifest.json
- */
-
 import path from 'path'
-import childProcess from 'child_process'
 import fs from 'fs-extra'
-import { logger } from '@ombro/logger'
+import execa from 'execa'
 import {
   Boilerplate,
   Lang,
   Manifest,
-  VRNBoilerplateConfig,
   verifyVRNBoilerplateConfig,
+  VRNBoilerplateConfig,
 } from '@vrn-deco/boilerplate-protocol'
 
 interface ScanPackage {
@@ -23,33 +16,7 @@ interface ScanPackage {
   path: string
 }
 
-if (require.main !== module) {
-  throw new Error('The module can only be used as a startup module.')
-}
-
-const PKG_DIR = path.resolve(__dirname, '..')
-
-logger.startLoading('Find boilerplate packages...')
-childProcess.exec('pnpm list -r --json --depth -1', { cwd: getRepoDir() }, (err, stdout) => {
-  if (err) throw err
-  try {
-    const pkgs = filterPackage(JSON.parse(stdout) as ScanPackage[])
-
-    logger.startLoading('Disposal data and archive...')
-    const manifest = archivePackage(pkgs)
-
-    logger.startLoading('Write vrn-manifest.json...')
-    writeManifest(manifest)
-
-    logger.stopLoading()
-    logger.done('Generated vrn-manifest.json.')
-  } catch (error) {
-    logger.error(error.stack)
-    process.exit(1)
-  }
-})
-
-function getRepoDir(): string {
+export function getRepoDir(): string {
   const repoDir = path.resolve(__dirname, '..', '..')
   const repoPkgPath = path.join(repoDir, 'package.json')
   const isValid =
@@ -58,13 +25,12 @@ function getRepoDir(): string {
   return repoDir
 }
 
-function filterPackage(pkgs: ScanPackage[]): ScanPackage[] {
-  return pkgs.filter((pkg) => /\/packages\/.+\/.+/.test(pkg.path))
+export function getAllBoilerplatePackage(): ScanPackage[] {
+  const { stdout } = execa.commandSync('pnpm list -r --json --depth -1', { cwd: getRepoDir() })
+  return JSON.parse(stdout).filter((pkg: ScanPackage) => /\/packages\/.+\/.+/.test(pkg.path))
 }
 
-type PKG = { name: string; version: string }
-
-function archivePackage(pkgs: ScanPackage[]): Manifest {
+export function archivePackage(pkgs: ScanPackage[]): Manifest {
   const langMap: Record<Lang['name'], Boilerplate[]> = {}
 
   pkgs.forEach(({ path: location }) => {
@@ -76,7 +42,7 @@ function archivePackage(pkgs: ScanPackage[]): Manifest {
     if (!fs.existsSync(pkgPath)) throw new Error(`${pkgPath} not exists`)
     if (!fs.existsSync(configPath)) throw new Error(`${configPath} not exists`)
 
-    const pkg = fs.readJsonSync(pkgPath) as PKG
+    const pkg = fs.readJsonSync(pkgPath) as { name: string; version: string }
     const config = fs.readJSONSync(configPath) as VRNBoilerplateConfig
 
     verifyVRNBoilerplateConfig(config, configPath)
@@ -86,7 +52,8 @@ function archivePackage(pkgs: ScanPackage[]): Manifest {
       desc: config.desc ?? '',
       package: pkg.name,
       version: pkg.version,
-      tags: config.tag ?? [],
+      tags: config.tags ?? [],
+      sort: config.sort ?? 100,
     }
     if (langMap[config.language]) {
       langMap[config.language]!.push(boilerplate)
@@ -104,8 +71,4 @@ function archivePackage(pkgs: ScanPackage[]): Manifest {
   })
 
   return manifest
-}
-
-function writeManifest(manifest: Manifest): void {
-  fs.writeJSONSync(path.join(PKG_DIR, 'vrn-manifest.json'), manifest, { spaces: 2 })
 }