+++ manifest/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @vrn-deco/boilerplate-manifest
 
+## 1.1.0
+
+### Minor Changes
+
+- Sorting the manifest
+
+## 1.0.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @vrn-deco/boilerplate-protocol@1.1.2
+
 ## 1.0.6
 
 ### Patch Changes